move emulation utils into a shared files; add ct nvfp4 scheme
